+++ src/Converter/Proxy/FromContext.php
@@ -1,0 +1,37 @@
+<?php
+declare(strict_types=1);
+
+namespace Smile\GdprDump\Converter\Proxy;
+
+use InvalidArgumentException;
+use Smile\GdprDump\Converter\ConverterInterface;
+use Smile\GdprDump\Converter\Helper\ArrayHelper;
+
+class FromContext implements ConverterInterface
+{
+    /**
+     * @var string
+     */
+    private $key;
+
+    /**
+     * @param array $parameters
+     */
+    public function __construct(array $parameters = [])
+    {
+        if (!array_key_exists('key', $parameters)) {
+            throw new InvalidArgumentException('The FromContext converter requires a "key" parameter.');
+        }
+
+        $this->key = (string) $parameters['key'];
+    }
+
+
+    /**
+     * @inheritdoc
+     */
+    public function convert($value, array $context = [])
+    {
+        return ArrayHelper::getPath($context, $this->key);
+    }
+}